feature-policy/ch-dpr-supported-by-feature-policy.tentative.html

<!DOCTYPE html>
<title>Test that ch-dpr is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://whatpr.org/html/3774/e32a6f8...ddb0544/infrastructure.html#policy-controlled-features">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('ch-dpr', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise ch-dpr.');
</script>
